+++ server/src/main/java/org/allaymc/server/block/component/crops/BlockCarrotsBaseComponentImpl.java
@@ -1,6 +1,7 @@
 package org.allaymc.server.block.component.crops;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.FortuneDropHelper;
 import org.allaymc.api.block.dto.BlockStateWithPos;
 import org.allaymc.api.block.property.type.BlockPropertyTypes;
 import org.allaymc.api.block.type.BlockType;
@@ -9,7 +10,6 @@ import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.type.ItemTypes;
 
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * @author daoge_cmd
@@ -22,10 +22,10 @@ public class BlockCarrotsBaseComponentImpl extends BlockCropsBaseComponentImpl {
     @Override
     public Set<ItemStack> getDrops(BlockStateWithPos blockState, ItemStack usedItem, Entity entity) {
         var growth = blockState.blockState().getPropertyValue(BlockPropertyTypes.GROWTH);
-        if (growth < 7) {
+        if (growth < BlockPropertyTypes.GROWTH.getMax()) {
             return Set.of(ItemTypes.CARROT.createItemStack());
         }
 
-        return Set.of(ItemTypes.CARROT.createItemStack(ThreadLocalRandom.current().nextInt(4) + 2));
+        return Set.of(ItemTypes.CARROT.createItemStack(FortuneDropHelper.binomial(usedItem, 1)));
     }
 }