+++ server/src/main/java/org/allaymc/server/block/component/grass/BlockShortGrassBaseComponentImpl.java
@@ -1,6 +1,7 @@
 package org.allaymc.server.block.component.grass;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.FortuneDropHelper;
 import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.dto.BlockStateWithPos;
 import org.allaymc.api.block.type.BlockType;
@@ -11,7 +12,6 @@ import org.allaymc.api.item.type.ItemTypes;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
 
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 
 import static org.allaymc.api.block.type.BlockTypes.*;
 
@@ -50,9 +50,8 @@ public class BlockShortGrassBaseComponentImpl extends BlockBaseComponentImpl {
 
     @Override
     public Set<ItemStack> getDrops(BlockStateWithPos blockState, ItemStack usedItem, Entity entity) {
-        var rand = ThreadLocalRandom.current();
-        if (rand.nextInt(8) == 0) {
-            return Set.of(ItemTypes.WHEAT_SEEDS.createItemStack(1));
+        if (FortuneDropHelper.bonusChanceDivisor(usedItem, 8, 2)) {
+            return Set.of(ItemTypes.WHEAT_SEEDS.createItemStack());
         }
         return Set.of(ItemAirStack.AIR_STACK);
     }